config sets up connection, db contains methods
